пути надо прописать

+++ src/main/java/web/dao/UserDao.java
@@ -6,8 +6,12 @@ import java.util.List;
 
 public interface UserDao {
     List<User> allUsers();
+
     void add(User user);
+
     void delete(User user);
+
     void edit(User user);
-    User getById(int id);
+
+    User getById(Long id);
 }
