+++ src/main/java/web/service/UserService.java
@@ -6,8 +6,12 @@ import java.util.List;
 
 public interface UserService {
     List<User> allUsers();
+
     void add(User user);
+
     void delete(User user);
+
     void edit(User user);
-    User getById(Integer id);
+
+    User getById(Long id);
 }
